✂️Refactors ViewModel
